Eliminated compiler warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2413 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csengine/covtree.cpp
@@ -969,11 +969,11 @@ again:
 
 void calc_size ()
 {
-printf ("1: %d (%dx%d)\n", sizeof(csCovTreeNode1), csCovTreeNode1::GetHorizontalSize(), csCovTreeNode1::GetVerticalSize());
-printf ("2: %d (%dx%d)\n", sizeof(csCovTreeNode2), csCovTreeNode2::GetHorizontalSize(), csCovTreeNode2::GetVerticalSize());
-printf ("3: %d (%dx%d)\n", sizeof(csCovTreeNode3), csCovTreeNode3::GetHorizontalSize(), csCovTreeNode3::GetVerticalSize());
-printf ("4: %d (%dx%d)\n", sizeof(csCovTreeNode4), csCovTreeNode4::GetHorizontalSize(), csCovTreeNode4::GetVerticalSize());
-printf ("5: %d (%dx%d)\n", sizeof(csCovTreeNode5), csCovTreeNode5::GetHorizontalSize(), csCovTreeNode5::GetVerticalSize());
+printf ("1: %lu (%dx%d)\n", (unsigned long)sizeof(csCovTreeNode1), csCovTreeNode1::GetHorizontalSize(), csCovTreeNode1::GetVerticalSize());
+printf ("2: %lu (%dx%d)\n", (unsigned long)sizeof(csCovTreeNode2), csCovTreeNode2::GetHorizontalSize(), csCovTreeNode2::GetVerticalSize());
+printf ("3: %lu (%dx%d)\n", (unsigned long)sizeof(csCovTreeNode3), csCovTreeNode3::GetHorizontalSize(), csCovTreeNode3::GetVerticalSize());
+printf ("4: %lu (%dx%d)\n", (unsigned long)sizeof(csCovTreeNode4), csCovTreeNode4::GetHorizontalSize(), csCovTreeNode4::GetVerticalSize());
+printf ("5: %lu (%dx%d)\n", (unsigned long)sizeof(csCovTreeNode5), csCovTreeNode5::GetHorizontalSize(), csCovTreeNode5::GetVerticalSize());
 }
 
 //---------------------------------------------------------------------------
